Switch CSPLayout to VF2Layout
In the Qiskit Terra 0.19.0 release a new pass VF2Layout was introduced
which models the search for a perfect layout as a subgraph isomorphism
problem that uses retworkx to model and find mappings. This was designed
to replace CSPLayout as a much faster alternative. The preset pass
managers in transpile() were updated to use this new pass as part of
0.20.0. This commit updates the pass manager used for mapping games to
use vf2layout instead of csplayout accordingly.