Remove undefined properties from VF2Layout initialization
In the previous commit we updated the mapping game pass manager to use
vf2layout. In that commit I copied the VF2Layout pass initialization
from the preset pass managers which have backend properties and/or a
target to provide more specifics about the backend target to the
compiler pass. This however doesn't exist for the mapping games and the
benchmarks are failing because of this error. This commit corrects the
oversight and removes the undefined fields.