Fix results storage handling in bishop
This commit fixes two small oversights in the results handling in
bishop. The first issues is the use of the non-existent .empty() method
on the benchmarks list. The parsed json outputs a list which does not
have an empty() method. This is corrected by just checking if the list
evals to False; if it's False then it's empty. The second issue is
switching the os.rename() function to shutil.move() function instead.
Both are in the stdlib and perform similar functions. But the syscall
that os.rename() uses only works when renaming a file within a
filesystem, while shutil.move() will move the file between filesystems
if necessary. Since the temp dir will be /tmp on most Linux systems and
that's sometimes a different filesystem (often a tmpfs) os.rename()
isn't a good choice.