Add tket queko mapping benchmark
This commit adds a new benchmark for testing tket's performance on the
queko benchmark suite. It runs queko with two different layout
(placement in the tket terminology) algorithms and the default
routing/mapping pass. The only place where these benchmarks are less
than ideal is that to get things to work correctly I needed to copy each
circuit in the timed function. Without this the inplace mutation and
shared references was causing the benchmarks to fail. Hopefully a better
long term solution can be found eventually.