Add config for pylint and black
This commit adds configuration to run black and pylint to establish
coding style and lint checking on the red queen repository. To do this a
tox.ini is added to automate venv creation and running both pylint and
black with consistent versions. A github actions CI configuration is
also added to ensure that any future proposed changes to the repo have
black tool's styling applied and complies with the pylint rules.