Pivot benchmark to be QASM based
This commit reworks the BV benchmark to run from a pre-defined QASM file
that is loaded as a qiskit circuit before benchmarking the compilation.
The previous version was just generating the circuit and qiskit and
running it directly. This wasn't ideal because we'll likely add
benchmarks for other compiler frameworks in the future and using qasm as
a standard input is useful for ensuring we're testing the same thing. To
generate the qasm you can run the game script directly which will update
the qasm files in the stored location.