Fix deprecated FakeMontreal import path
In qiskit-terra 0.21.0 the qiskit.test.mock path was deprecated and the
code was migrated to live in qiskit.providers.fake_provider. However,
while the old path still exists it works by re-exporting the new path
from the old path. Pylint is able to realize this and is throwing an
error about the FakeMontreal name not being found in the
qiskit.test.mock module. To fix this lint failure this commit just
updates the import to use the new path to avoid this issue. We would
have to make this change anyway after the deprecation cycle is finished
and qiskit.test.mock is removed.