Add Bernstein Vazirani circuit benchmark for Qiskit (#6)
* Add Bernstein Vazirani circuit benchmark for Qiskit

This commit adds a new benchmark for bernstein vazirani circuit
compilation comparing the output circuit at different optimization
levels with Qiskit's compiler and also comparing the noisy simulation
fidelity against an ideal result.

* Pivot benchmark to be QASM based

This commit reworks the BV benchmark to run from a pre-defined QASM file
that is loaded as a qiskit circuit before benchmarking the compilation.
The previous version was just generating the circuit and qiskit and
running it directly. This wasn't ideal because we'll likely add
benchmarks for other compiler frameworks in the future and using qasm as
a standard input is useful for ensuring we're testing the same thing. To
generate the qasm you can run the game script directly which will update
the qasm files in the stored location.

* Update mid-circuit measurement case to reset ancilla too

Co-authored-by: Paul Nation <nonhermitian@gmail.com>

* Fix lint

Co-authored-by: Paul Nation <nonhermitian@gmail.com>