Reorganize Red Queen into a python package
This commit reorganize the red queen project into a python project.
Previously red queen wasn't organized as a single package which caused
potential issues when trying to run or interact between the packages.
This commit changes the repository structure to move all the code into
the red_queen package so it is all self contained. This also enables us
to create a python package with setuptools which increases future
flexibility by enabling us to publish console scripts which will be
include when the package is built and installed.