Fix hidden shift benchmark
In the recently merged hidden shift benchmark, which was added as part
of #20, there was n oversight that slipped through as part of some code
review refactoring. An extra parameterized argument which was removed in
an earlier revision of the PR was left on the function signature. This
causes the benchmark to fail to execute because pytest doesn't know how
to populate that parameter as it no longer exists. This commit corrects
this oversight and removes the leftover argument. Hopefully in the
future we'll have #29 which will catch these easy to miss issues
pre-merge and block us from merging broken code.