Fixed bash syntax in action.yml file for changed-benchmark (#21)
There was an error in the `action.yml` file that caused the files-changed condition to ignore any new python files. 
This was due to an error in string parsing that was returning an empty string in every case instead of returning the string without the `__init__.py`
The error was fixed by making the `steps.changed-benchmarks.outputs.all_changed_and_modified_files` a string stored in a variable.